Update README.md
Added examples of how to determine when the RentDynamics SDK is returning at error. I would recommend normalizing the response, meaning always return a JSON object regardless of error or not in the SDK, but this would be a breaking change for any current users. I would be happy to put up a PR if it's something you all are interested in.